creating voucher
