+++ frontend-contracts/newconfig.js
@@ -1,5 +1,5 @@
 
-  export const nftaddress = "0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512"
+  export const nftaddress = "0x5FbDB2315678afecb367f032d93F642f64180aa3"
   export const redeemeraddress = "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266"
   export const minteraddress = "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"
