+++ frontend-contracts/pages/_app.js
@@ -7,8 +7,13 @@ function MyApp({ Component, pageProps }) {
       <nav className="border-b p-6">
         <p className="text-4xl font-bold">Testing</p>
         <div className="flex mt-4"> 
+        {/*
           <Link href="/">
             <a className="mr-6 text-blue-500">Home</a>
+          </Link>
+        */}
+          <Link href="/">
+            <a href="url" className="mr-6 text-blue-500">Marketplace</a>
           </Link>
           <Link href="/create-item">
             <a className="mr-6 text-blue-500">Mint and Sell</a>
@@ -19,9 +24,7 @@ function MyApp({ Component, pageProps }) {
           <Link href="/creator-dashboard">
             <a className="mr-6 text-blue-500">Creator Dashboard</a>
           </Link>
-          <Link href="/listVouchers">
-            <a href="url" className="mr-6 text-blue-500">Vouchers</a>
-          </Link>
+         
         </div>
       </nav>
       <Component {...pageProps} />
